Seperated code into diff modules, added transition between login/sign up, removed old code and put dead code in a seperate place.
